+++ tests/test_device.py
@@ -72,9 +72,9 @@ class TestDeviceIntegration:
     @pytest.mark.parametrize("d", shortnames)
     def test_load_device(self, d):
         """Test that the device loads correctly"""
-        dev = qml.device(d, wires=2, shots=1024)
+        dev = qml.device(d, wires=2)
         assert dev.num_wires == 2
-        assert dev.shots.total_shots == 1024
+        assert dev.shots.total_shots is None
         assert dev.short_name == d
 
     @pytest.mark.parametrize("d", shortnames)
@@ -83,9 +83,18 @@ class TestDeviceIntegration:
         with pytest.raises(TypeError, match="missing 1 required positional argument"):
             qml.device(d)
 
-        # IonQ devices do not allow shots=None
+        # IonQ devices allow shots=None
+        dev = qml.device(d, wires=1, shots=None)
+
+        # But the execution will raise error
+        @qml.qnode(dev)
+        def circuit():
+            """Reference QNode"""
+            qml.PauliX(wires=0)
+            return qml.expval(qml.PauliZ(0))
+
         with pytest.raises(ValueError, match="does not support analytic"):
-            qml.device(d, wires=1, shots=None)
+            circuit()
 
     def test_emptycircuit_warning(self, mocker):
         """Test warning raised on submission of an empty circuit."""
@@ -183,12 +192,13 @@ class TestDeviceIntegration:
     @pytest.mark.parametrize("shots", [8192])
     def test_one_qubit_circuit(self, shots, requires_api, tol):
         """Test that devices provide correct result for a simple circuit"""
-        dev = qml.device("ionq.simulator", wires=1, shots=shots)
+        dev = qml.device("ionq.simulator", wires=1)
 
         a = 0.543
         b = 0.123
         c = qml.numpy.array(0.987, requires_grad=False)
 
+        @qml.set_shots(shots)
         @qml.qnode(dev)
         def circuit(x, y, z):
             """Reference QNode"""
@@ -202,8 +212,9 @@ class TestDeviceIntegration:
     @pytest.mark.parametrize("shots", [8192])
     def test_one_qubit_ordering(self, shots, requires_api, tol):
         """Test that probabilities are returned with the correct qubit ordering"""
-        dev = qml.device("ionq.simulator", wires=2, shots=shots)
+        dev = qml.device("ionq.simulator", wires=2)
 
+        @qml.set_shots(shots)
         @qml.qnode(dev)
         def circuit():
             qml.PauliX(wires=1)
@@ -216,11 +227,19 @@ class TestDeviceIntegration:
     def test_prob_no_results(self, d):
         """Test that the prob attribute is
         None if no job has yet been run."""
-        dev = qml.device(d, wires=1, shots=1)
+        dev = qml.device(d, wires=1)
         assert dev.prob is None
 
     @pytest.mark.parametrize(
-        "backend", ["aria-1", "aria-2", "forte-1", "forte-enterprise-1", "forte-enterprise-2", None]
+        "backend",
+        [
+            "aria-1",
+            "aria-2",
+            "forte-1",
+            "forte-enterprise-1",
+            "forte-enterprise-2",
+            None,
+        ],
     )
     def test_backend_initialization(self, backend):
         """Test that the device initializes with the correct backend."""
@@ -233,7 +252,7 @@ class TestDeviceIntegration:
 
     def test_recording_when_pennylane_tracker_active(self, requires_api):
         """Test recording device execution history via pennnylane tracker class."""
-        dev = SimulatorDevice(wires=(0,), gateset="native")
+        dev = SimulatorDevice(wires=(0,), gateset="native", shots=1024)
         dev.tracker = qml.Tracker()
         dev.tracker.active = True
         dev.tracker.reset()
@@ -256,7 +275,7 @@ class TestDeviceIntegration:
 
     def test_not_recording_when_pennylane_tracker_not_active(self, requires_api):
         """Test recording device not executed when tracker is inactive."""
-        dev = SimulatorDevice(wires=(0,), gateset="native")
+        dev = SimulatorDevice(wires=(0,), gateset="native", shots=1024)
         dev.tracker = qml.Tracker()
         dev.tracker.active = False
         dev.tracker.reset()
@@ -268,7 +287,7 @@ class TestDeviceIntegration:
 
     def test_warning_on_empty_circuit(self, requires_api):
         """Test warning are shown when circuit is empty."""
-        dev = SimulatorDevice(wires=(0,), gateset="native")
+        dev = SimulatorDevice(wires=(0,), gateset="native", shots=1024)
         with qml.tape.QuantumTape(shots=1024) as tape1:
             qml.probs(wires=[0])
         with pytest.warns(
@@ -280,10 +299,13 @@ class TestDeviceIntegration:
     @mock.patch("logging.Logger.isEnabledFor", return_value=True)
     @mock.patch("logging.Logger.debug")
     def test_batch_execute_logging_when_enabled(
-        self, mock_logging_debug_method, mock_logging_is_enabled_for_method, requires_api
+        self,
+        mock_logging_debug_method,
+        mock_logging_is_enabled_for_method,
+        requires_api,
     ):
         """Test logging invoked in batch_execute method."""
-        dev = SimulatorDevice(wires=(0,), gateset="native")
+        dev = SimulatorDevice(wires=(0,), gateset="native", shots=1024)
         with qml.tape.QuantumTape(shots=1024) as tape1:
             GPI(0, wires=[0])
             qml.probs(wires=[0])
@@ -296,7 +318,7 @@ class TestDeviceIntegration:
         """Test invoking probability() method raises exception if circuit index not
         previously set when multiple circuits are submitted in one job.
         """
-        dev = SimulatorDevice(wires=(0,), gateset="native")
+        dev = SimulatorDevice(wires=(0,), gateset="native", shots=1024)
         with qml.tape.QuantumTape(shots=1024) as tape1:
             GPI(0.5, wires=[0])
             qml.probs(wires=[0])
@@ -310,7 +332,7 @@ you want to access must be first set via the set_current_circuit_index device me
 
     def test_batch_execute_probabilities(self, requires_api):
         """Test batch_execute method when computing circuit probabilities."""
-        dev = SimulatorDevice(wires=(0, 1, 2), gateset="native")
+        dev = SimulatorDevice(wires=(0, 1, 2), gateset="native", shots=1024)
         with qml.tape.QuantumTape(shots=1024) as tape1:
             GPI(0.5, wires=[0])
             GPI2(0, wires=[1])
@@ -337,7 +359,7 @@ you want to access must be first set via the set_current_circuit_index device me
 
     def test_batch_execute_probabilities_with_shot_vector(self, requires_api):
         """Test batch_execute method with shot vector."""
-        dev = SimulatorDevice(wires=(0, 1, 2), gateset="native")
+        dev = SimulatorDevice(wires=(0, 1, 2), gateset="native", shots=1024)
         dev._shot_vector = (ShotCopies(1, 3),)
         with qml.tape.QuantumTape(shots=1024) as tape1:
             GPI(0, wires=[0])
@@ -347,7 +369,7 @@ you want to access must be first set via the set_current_circuit_index device me
 
     def test_batch_execute_variance(self, requires_api):
         """Test batch_execute method when computing variance of an observable."""
-        dev = SimulatorDevice(wires=(0,), gateset="native")
+        dev = SimulatorDevice(wires=(0,), gateset="native", shots=1024)
         with qml.tape.QuantumTape(shots=1024) as tape1:
             GPI(0, wires=[0])
             qml.var(qml.PauliZ(0))
@@ -360,7 +382,7 @@ you want to access must be first set via the set_current_circuit_index device me
 
     def test_batch_execute_expectation_value(self, requires_api):
         """Test batch_execute method when computing expectation value of an observable."""
-        dev = SimulatorDevice(wires=(0,), gateset="native")
+        dev = SimulatorDevice(wires=(0,), gateset="native", shots=1024)
         with qml.tape.QuantumTape(shots=1024) as tape1:
             GPI(0, wires=[0])
             qml.expval(qml.PauliZ(0))
@@ -374,7 +396,7 @@ you want to access must be first set via the set_current_circuit_index device me
     def test_batch_execute_expectation_value_with_diagonalization_rotations(self, requires_api):
         """Test batch_execute method when computing expectation value of an
         observable that requires rotations for diagonalization."""
-        dev = SimulatorDevice(wires=(0,), gateset="qis")
+        dev = SimulatorDevice(wires=(0,), gateset="qis", shots=1024)
         with qml.tape.QuantumTape(shots=1024) as tape1:
             qml.Hadamard(0)
             qml.expval(qml.PauliX(0))
@@ -388,7 +410,7 @@ you want to access must be first set via the set_current_circuit_index device me
         """Test invoking prob device property raises exception if circuit index not
         previously set when multiple circuits are submitted in one job.
         """
-        dev = SimulatorDevice(wires=(0,), gateset="native")
+        dev = SimulatorDevice(wires=(0,), gateset="native", shots=1024)
         with qml.tape.QuantumTape(shots=1024) as tape1:
             GPI(0, wires=[0])
             qml.probs(wires=[0])
@@ -405,7 +427,7 @@ you want to access must be first set via the set_current_circuit_index device me
 
     def test_batch_execute_prob_property(self, requires_api):
         """Test batch_execute method with invoking invoking prob device property."""
-        dev = SimulatorDevice(wires=(0,), gateset="native")
+        dev = SimulatorDevice(wires=(0,), gateset="native", shots=1024)
         with qml.tape.QuantumTape(shots=1024) as tape1:
             GPI(0, wires=[0])
             qml.sample(qml.PauliZ(0))
@@ -422,7 +444,7 @@ you want to access must be first set via the set_current_circuit_index device me
 
     def test_batch_execute_counts(self, requires_api):
         """Test batch_execute method when computing counts."""
-        dev = SimulatorDevice(wires=(0,), gateset="native")
+        dev = SimulatorDevice(wires=(0,), gateset="native", shots=1024)
         with qml.tape.QuantumTape(shots=1024) as tape1:
             GPI(0, wires=[0])
             qml.counts(qml.PauliZ(0))
@@ -453,7 +475,7 @@ you want to access must be first set via the set_current_circuit_index device me
             def __copy__(self):
                 return CountState(state=self.state)
 
-        dev = SimulatorDevice(wires=(0,), gateset="native")
+        dev = SimulatorDevice(wires=(0,), gateset="native", shots=1024)
 
         with qml.tape.QuantumTape(shots=1024) as tape1:
             GPI(0, wires=[0])
@@ -478,7 +500,7 @@ class TestJobAttribute:
             pass
 
         mocker.patch("pennylane_ionq.device.IonQDevice._submit_job", mock_submit_job)
-        dev = IonQDevice(wires=(0,), target="foo")
+        dev = IonQDevice(wires=(0,), target="foo", shots=1024)
 
         with qml.tape.QuantumTape() as tape:
             qml.PauliX(0)
@@ -503,7 +525,7 @@ class TestJobAttribute:
             pass
 
         mocker.patch("pennylane_ionq.device.IonQDevice._submit_job", mock_submit_job)
-        dev = IonQDevice(wires=(0,), target="foo")
+        dev = IonQDevice(wires=(0,), target="foo", shots=1024)
 
         with qml.tape.QuantumTape() as tape:
             qml.PauliX(0)
@@ -560,7 +582,7 @@ class TestJobAttribute:
             pass
 
         mocker.patch("pennylane_ionq.device.IonQDevice._submit_job", mock_submit_job)
-        dev = IonQDevice(wires=(0,))
+        dev = IonQDevice(wires=(0,), shots=1024)
 
         with qml.tape.QuantumTape() as tape:
             qml.RX(1.2345, wires=0)
@@ -592,7 +614,7 @@ class TestJobAttribute:
             pass
 
         mocker.patch("pennylane_ionq.device.IonQDevice._submit_job", mock_submit_job)
-        dev = IonQDevice(wires=(0, 1, 2), gateset="native")
+        dev = IonQDevice(wires=(0, 1, 2), gateset="native", shots=1024)
 
         with qml.tape.QuantumTape() as tape:
             GPI(0.1, wires=[0])
@@ -640,7 +662,7 @@ class TestJobAttribute:
             raise StopExecute()
 
         mocker.patch("pennylane_ionq.device.SimulatorDevice._submit_job", mock_submit_job)
-        dev = SimulatorDevice(wires=(0,), gateset="native")
+        dev = SimulatorDevice(wires=(0,), gateset="native", shots=1024)
 
         with qml.tape.QuantumTape(shots=1024) as tape1:
             GPI(0.7, wires=[0])
@@ -718,15 +740,15 @@ class TestJobAttribute:
         """Test SWAP gate operation is correctly processed and sent to IonQ."""
         dev = qml.device("ionq.simulator", wires=2, gateset="qis")
 
-        with qml.tape.QuantumTape() as tape:
+        with qml.tape.QuantumTape(shots=1024) as tape:
             qml.PauliX(wires=0)
             qml.SWAP(wires=[0, 1])
             qml.probs(wires=[0, 1])
 
-        result_ionq = dev.batch_execute([tape])
+        result_ionq = dev.execute([tape])
 
         simulator = qml.device("default.qubit", wires=2)
-        result_simulator = qml.execute([tape], simulator)
+        result_simulator = qml.execute([tape.copy(shots=None)], simulator)
 
         assert np.allclose(
             result_ionq, result_simulator, atol=1e-5
@@ -736,15 +758,15 @@ class TestJobAttribute:
         """Test a controlled gate operation is correctly processed and sent to IonQ."""
         dev = qml.device("ionq.simulator", wires=2, gateset="qis")
 
-        with qml.tape.QuantumTape() as tape:
+        with qml.tape.QuantumTape(shots=1024) as tape:
             qml.Hadamard(wires=0)
             qml.CNOT(wires=[0, 1])
             qml.probs(wires=[0, 1])
 
-        result_ionq = dev.batch_execute([tape])
+        result_ionq = dev.execute([tape])
 
         simulator = qml.device("default.qubit", wires=2)
-        result_simulator = qml.execute([tape], simulator)
+        result_simulator = qml.execute([tape.copy(shots=None)], simulator)
 
         assert np.allclose(
             result_ionq, result_simulator, atol=1e-5