+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 
 ### Breaking changes 💔
 
+* Changed the default `shots` to `None` from `1024` for all devices. Additionally, shots validation now
+  happens at execution.
+  [(#149)](https://github.com/PennyLaneAI/PennyLane-IonQ/pull/149)
+
 * Remove support for Python 3.10 and add support for Python 3.13.
   [(#144)](https://github.com/PennyLaneAI/PennyLane-IonQ/pull/144)
 
@@ -22,6 +26,7 @@
 
 This release contains contributions from (in alphabetical order):
 
+Yushao Chen,
 Radu Marginean,
 Andrija Paurevic.